Refactor RB unit tests
